backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
gunicorn==21.2.0
python-dotenv==0.21.1
sqlalchemy==1.4.41
pydantic==1.10.7
docker==6.1.2
aiohttp==3.8.4
python-multipart==0.0.6
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
alembic==1.11.1
prometheus-client==0.17.1
kubernetes==26.1.0 